Updating navigation handling.
