+++ client/src/App.jsx
@@ -7,7 +7,7 @@ import {
 } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
 
-import Nav from './components/Nav';
+import Nav from './components/navigation.jsx';
 import Footer from './components/ui/footer.jsx';
 
 const httpLink = createHttpLink({
